backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1
python-socketio==5.11.4
python-engineio==4.10.1
celery==5.4.0
pydantic==2.10.3
python-dotenv==1.0.1
greenlet==3.1.1
groq==1.0.0
